fix: use router for navigation
* Move AuthProvider to top level to avoid remounts when navigating
* Remove Main.tsx component
* MUI buttons need to be injected with NavLink component in order to use
  router (oops)
* update login page to wait to redirect
* update user page to use nav component upon no user
