+++ src/components/SignIn.tsx
@@ -1,23 +1,29 @@
 import {Typography, Box, Button, TextField} from '@mui/material';
-import React from 'react';
+import React, {useEffect} from 'react';
 import {useNavigate} from 'react-router-dom';
 import useForm from '../hooks/useForm';
 import {useAuth} from '../provider/AuthProvider';
 
-// This is the auth'd page.
 function SignIn() {
-  const {signIn} = useAuth();
+  const {signIn, user} = useAuth();
   const navigate = useNavigate();
   const {values, errors, handleChange, handleSubmit} = useForm(doLogin);
+
   function doLogin() {
     signIn(values['email'], values['pass'])
       .then(() => {
-        navigate('/userpage');
+        console.log('User logged in.');
       })
       .catch(error => {
         console.error(`${error.code}: ${error.message}`);
       });
   }
+
+  useEffect(() => {
+    if (user) {
+      navigate('/userpage');
+    }
+  }, [user]);
 
   return (
     <>
