+++ src/components/UserPage.tsx
@@ -1,46 +1,46 @@
 import React, {useState, useEffect} from 'react';
 import {Typography} from '@mui/material';
-import {getAuth, onAuthStateChanged} from 'firebase/auth';
-import {useNavigate} from 'react-router-dom';
+import {useAuth} from '../provider/AuthProvider';
+import {Navigate} from 'react-router-dom';
 import {getStorage, getDownloadURL, ref} from 'firebase/storage';
 
 function UserPage() {
-  const [user, setUser] = useState<firebase.User | null>(null);
-  const [image, setImage] = useState<String | null>(null);
-  const auth = getAuth();
-  const navigate = useNavigate();
+  const {user} = useAuth();
+  const [image, setImage] = useState<string | null>(null);
   const storage = getStorage();
 
   useEffect(() => {
-    onAuthStateChanged(auth, firebaseuser => {
-      if (firebaseuser) {
-        setUser(firebaseuser);
-        const gsRef = ref(
-          storage,
-          `gs://authflow-89fd5.appspot.com/${firebaseuser.uid}`
-        );
-        getDownloadURL(gsRef)
-          .then(prom => {
-            setImage(prom);
-          })
-          .catch(error => {
-            console.error(`${error}`);
-          });
-      } else {
-        navigate('/');
-      }
-    });
+    if (user) {
+      const gsRef = ref(storage, `gs://authflow-89fd5.appspot.com/${user.uid}`);
+      getDownloadURL(gsRef)
+        .then(prom => {
+          setImage(prom);
+        })
+        .catch(error => {
+          console.error(error.message);
+        });
+    } else {
+      console.log('User:', user);
+    }
   }, []);
 
   return (
     <>
-      <Typography>User Info. Email: {user?.email}</Typography>
-      {image && (
-        <img
-          style={{maxWidth: 300}}
-          alt={`${user?.email}'s avatar`}
-          src={image}
-        />
+      {user ? (
+        <>
+          <Typography>User Info.Email: {user?.email}</Typography>
+          {image ? (
+            <img
+              style={{maxWidth: 300}}
+              alt={`${user?.email}'s avatar`}
+              src={image}
+            />
+          ) : (
+            <Typography>No image found.</Typography>
+          )}
+        </>
+      ) : (
+        <Navigate to="/" />
       )}
     </>
   );
